Fix options validation structure
