+++ packages/lib/src/validation/index.ts
@@ -1,3 +1,4 @@
 export * from './base';
 export * from './commands';
+export * from './options';
 export * from './checks';
